Add a cast to quiet a compiler warning.
--HG--
branch : 1.7

+++ iolog.c
@@ -289,7 +289,7 @@ io_log_open()
     gettimeofday(&last_time, NULL);
 
     /* XXX - log more stuff?  window size? environment? */
-    fprintf(io_logfile, "%ld:%s:%s:%s:%s\n", last_time.tv_sec, user_name,
+    fprintf(io_logfile, "%ld:%s:%s:%s:%s\n", (long)last_time.tv_sec, user_name,
         runas_pw->pw_name, runas_gr ? runas_gr->gr_name : "", user_tty);
     fprintf(io_logfile, "%s\n", user_cwd);
     fprintf(io_logfile, "%s%s%s\n", user_cmnd, user_args ? " " : "",
